move blockchain specific logic from settings to blockchain model
