+++ src/app/udapp/run-tab.js
@@ -9,7 +9,7 @@ const EventManager = require('../../lib/events')
 const Card = require('../ui/card')
 
 const css = require('../tabs/styles/run-tab-styles')
-const Settings = require('../tabs/runTab/model/settings.js')
+// const Settings = require('../tabs/runTab/model/settings.js')
 const SettingsUI = require('../tabs/runTab/settings.js')
 const Recorder = require('../tabs/runTab/model/recorder.js')
 const RecorderUI = require('../tabs/runTab/recorder.js')
@@ -124,8 +124,8 @@ export class RunTab extends LibraryPlugin {
   }
 
   renderSettings (udapp) {
-    var settings = new Settings(this.executionContext, udapp)
-    this.settingsUI = new SettingsUI(settings, this.networkModule)
+    // var settings = new Settings(this.blockchain, this.executionContext, udapp)
+    this.settingsUI = new SettingsUI(this.blockchain, this.networkModule)
 
     this.settingsUI.event.register('clearInstance', () => {
       this.event.trigger('clearInstance', [])
